Move git to Recommends in kali-linux-core
The idea is that one can have a slimmer "core" if you install
kali-linux-core with --no-install-recommends. For instance git pulls
more than 110Mb of dependencies on a bare docker image...

But we put git into Depends of kali-linux-headless so that it
gets installed with that package even when the core was installed in a
minimalistic way.