kali-desktop-core: Recommends += orca
orca adds accessibility for users with visual impairment.

In Debian, orca is recommended for most desktops: xfce, gnome, kde,
lxde, lxqt, mate, cinnamon (see 'apt rdepends orca'). For Kali, let's
make it simple and add the recommendation to kali-desktop-core, so that
it's available in all desktops.

It means that we also install it for the packages kali-desktop-{i3,e17}
where it's not clear whether orca is really supported or integrated. But
I guess it does no harm either.

Closes: #10