Add Conflicts: wireplumber, Recommends: pipewire-media-session
In a recent change in the pipewire package in Debian, wireplumber was
introduced as a Recommends. Unfortunately it was added BEFORE
pipewire-media-session, and that broke the sound for many users.

Cf:
- https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=999363
- https://salsa.debian.org/utopia-team/pipewire/-/commit/bf12b0177

Even though the pipewire package is fixed from 0.3.40-2 onward, systems
which already have wireplumber installed due to this change are still
broken. Running `apt upgrade` won't remove wireplumber and replace it by
pipewire-media-session.

And that's pretty much every Kali system, since Kali 2021.4 was shipped
with the broken pipewire...

So we must force the removal of wireplumber, and make sure it's replaced
with pipewire-media-session. It seems that the best way to do that is to
add a (temporary) Conflicts+Recommends. A Conflicts alone is not enough.