Recommend fuse3 in kali-linux-core & kali-linux-headless
The command fuse is either provided by the package fuse (aka. fuse2),
either by the package fuse3. What we get in Kali depends on the packages
installed, for some desktops it will be fuse3, for others (or for
headless) it will be fuse. In short, the rule is that as long as no
package requires fuse3 explicitly, fuse is likely to be installed.

With this commit, we try to make sure that all Kali installations come
with fuse3.

Apart from bringing consistency, it should also solve an issue with the
Kali Linux installer, open-vm-tools and some desktops (KDE apparently).
Indeed, if fuse is installed when the bulk of the packages are
installed, when the desktop is installed, then the installer can't
install open-vm-tools, as it requires fuse3, but at this stage of the
installation, the installer refuses to remove packages, so it can't
replace fuse by fuse3.

This is the case for KDE at the moment.

As for kali-linux-headless: both curlftpfs and testdisk depend on fuse,
and no package depends on fuse3, so fuse wins.

So we recommend fuse in kali-desktop-core and kali-linux-headless

For even more details, refer to:
https://gitlab.com/kalilinux/packages/kali-meta/-/issues/15

Fixes: #15